Extract existential, environmental, and nominal predication
